+++ packages/core/src/entity/EditorModeEntityStore.ts
@@ -2,10 +2,24 @@ import { EditorEntityStore, RequestedEntitiesMessage } from '@contentful/visual-
 import type { Asset, AssetFile, Entry, UnresolvedLink } from 'contentful';
 import { sendMessage } from '../communication/sendMessage';
 
+// The default of 3s in the EditorEntityStore is sometimes timing out and
+// leads to not rendering bound content and design components.
+const REQUEST_TIMEOUT = 10000;
+
 export class EditorModeEntityStore extends EditorEntityStore {
+  public locale: string;
+
   constructor({ entities, locale }: { entities: Array<Entry | Asset>; locale: string }) {
+    console.debug(
+      `[exp-builder.sdk] Initializing editor entity store with ${entities.length} entities for locale ${locale}.`,
+      { entities }
+    );
+
     const subscribe = (method: unknown, cb: (payload: RequestedEntitiesMessage) => void) => {
       const listeners = (event: MessageEvent) => {
+        if (typeof event.data !== 'string') {
+          return;
+        }
         const data: {
           source: 'composability-app';
           eventType: string;
@@ -13,6 +27,7 @@ export class EditorModeEntityStore extends EditorEntityStore {
         } = JSON.parse(event.data);
 
         if (typeof data !== 'object' || !data) return;
+
         if (data.source !== 'composability-app') return;
         if (data.eventType === method) {
           cb(data.payload);
@@ -30,20 +45,33 @@ export class EditorModeEntityStore extends EditorEntityStore {
       };
     };
 
-    super({ entities, sendMessage, subscribe, locale });
+    super({ entities, sendMessage, subscribe, locale, timeoutDuration: REQUEST_TIMEOUT });
+    this.locale = locale;
   }
+  /**
+   * This function collects and returns the list of requested entries and assets. Additionally, it checks
+   * upfront whether any async fetching logic is actually happening. If not, it returns a plain `false` value, so we
+   * can detect this early and avoid unnecessary re-renders.
+   * @param entityLinks
+   * @returns false if no async fetching is happening, otherwise a promise that resolves when all entities are fetched
+   */
+  fetchEntities(entityLinks: UnresolvedLink<'Entry' | 'Asset'>[]): false | Promise<void> {
+    const entryLinks = entityLinks.filter((link) => link.sys?.linkType === 'Entry');
+    const assetLinks = entityLinks.filter((link) => link.sys?.linkType === 'Asset');
 
-  async fetchEntities(entityLinks: UnresolvedLink<'Entry' | 'Asset'>[]) {
-    const entryLinks = entityLinks.filter((link) => link.sys.linkType === 'Entry');
-    const assetLinks = entityLinks.filter((link) => link.sys.linkType === 'Asset');
+    const uniqueEntryIds = [...new Set(entryLinks.map((link) => link.sys.id))];
+    const uniqueAssetIds = [...new Set(assetLinks.map((link) => link.sys.id))];
 
-    const uniqueEntryLinks = new Set(entryLinks.map((link) => link.sys.id));
-    const uniqueAssetLinks = new Set(assetLinks.map((link) => link.sys.id));
+    const { missing: missingEntryIds } = this.getEntitiesFromMap('Entry', uniqueEntryIds);
+    const { missing: missingAssetIds } = this.getEntitiesFromMap('Asset', uniqueAssetIds);
 
-    return await Promise.allSettled([
-      this.fetchEntries([...uniqueEntryLinks]),
-      this.fetchAssets([...uniqueAssetLinks]),
-    ]);
+    // Return false to indicate that no async fetching is happening
+    if (!missingAssetIds.length && !missingEntryIds.length) return false;
+
+    // Entries and assets will be stored in entryMap and assetMap
+    return Promise.all([this.fetchEntries(uniqueEntryIds), this.fetchAssets(uniqueAssetIds)]).then(
+      () => Promise.resolve()
+    );
   }
 
   getValue(