+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from './utils';
-export * from './definitions/styles';
+export * from './definitions';
 export * from './entity';
 export * from './communication';
 export * from './enums';
+export * from './registries';