+++ packages/core/src/utils/index.ts
@@ -5,4 +5,5 @@ export * from './transformers';
 export * from './utils';
 export * from './typeguards';
 export * from './supportedModes';
-export * from './getValueForBreakpoint';
+export * from './breakpoints';
+export * from './isLinkToAsset';