+++ packages/experience-builder-sdk/src/core/componentRegistry.ts
@@ -10,11 +10,11 @@ import {
   CONTENTFUL_SECTION_ID,
 } from '@contentful/experience-builder-core/constants';
 import { ContentfulContainer } from '@contentful/experience-builder-components';
-import { containerDefinition } from './definitions/components';
 import {
   builtInStyles as builtInStyleDefinitions,
   optionalBuiltInStyles,
   sendMessage,
+  containerDefinition,
 } from '@contentful/experience-builder-core';
 import { withComponentWrapper } from '../utils/withComponentWrapper';
 import { SDK_VERSION } from '../constants';