+++ packages/visual-editor/src/hooks/useComponent.ts
@@ -1,26 +1,47 @@
-import { useEditorStore } from '@/store/editor';
-import {
+import type {
   ComponentRegistration,
   CompositionComponentNode,
+  ResolveDesignValueType,
 } from '@contentful/experience-builder-core/types';
 import { useMemo } from 'react';
-import { ResolveDesignValueType } from './useBreakpoints';
 import { useComponentProps } from './useComponentProps';
 import { builtInComponents } from '@/types/constants';
+import { DESIGN_COMPONENT_NODE_TYPE } from '@contentful/experience-builder-core/constants';
+import { DesignComponent } from '@contentful/experience-builder-components';
+import { resolveDesignComponent } from '@/utils/designComponentUtils';
+import { componentRegistry, createDesignComponentRegistration } from '@/store/registries';
+import { useEntityStore } from '@/store/entityStore';
 
 interface ComponentParams {
   node: CompositionComponentNode;
   resolveDesignValue: ResolveDesignValueType;
-  areEntitiesFetched: boolean;
 }
 
-export const useComponent = ({ node, resolveDesignValue, areEntitiesFetched }: ComponentParams) => {
-  const componentRegistry = useEditorStore((state) => state.componentRegistry);
+export const useComponent = ({ node: rawNode, resolveDesignValue }: ComponentParams) => {
+  const areEntitiesFetched = useEntityStore((state) => state.areEntitiesFetched);
+  const entityStore = useEntityStore((state) => state.entityStore);
+
+  const node = useMemo(() => {
+    if (rawNode.type === DESIGN_COMPONENT_NODE_TYPE && areEntitiesFetched) {
+      return resolveDesignComponent({
+        node: rawNode,
+        entityStore,
+      });
+    }
+
+    return rawNode;
+  }, [areEntitiesFetched, rawNode, entityStore]);
 
   const componentRegistration = useMemo(() => {
-    const id = node.data.blockId as string;
+    const registration = componentRegistry.get(node.data.blockId as string);
 
-    return componentRegistry.get(id) as ComponentRegistration;
+    if (node.type === DESIGN_COMPONENT_NODE_TYPE && !registration) {
+      return createDesignComponentRegistration({
+        definitionId: node.data.blockId as string,
+        component: DesignComponent,
+      }) as ComponentRegistration;
+    }
+    return registration as ComponentRegistration;
   }, [node]);
 
   const componentId = node.data.id;
@@ -35,6 +56,7 @@ export const useComponent = ({ node, resolveDesignValue, areEntitiesFetched }: C
   const Component = builtInComponents[node.data.blockId!] || componentRegistration.component;
 
   return {
+    node,
     componentId,
     Component,
     props,