+++ packages/visual-editor/src/hooks/useComponentProps.tsx
@@ -4,26 +4,27 @@ import {
   buildCfStyles,
   calculateNodeDefaultHeight,
   transformContentValue,
+  isLinkToAsset,
 } from '@contentful/experience-builder-core';
 import {
   CONTENTFUL_CONTAINER_ID,
   CF_STYLE_ATTRIBUTES,
+  DESIGN_COMPONENT_NODE_TYPE,
 } from '@contentful/experience-builder-core/constants';
 import type {
   StyleProps,
   CompositionVariableValueType,
   CompositionComponentNode,
+  ResolveDesignValueType,
   ComponentRegistration,
+  Link,
 } from '@contentful/experience-builder-core/types';
 import { useMemo } from 'react';
-import { useShallow } from 'zustand/react/shallow';
 import { useStyleTag } from './useStyleTag';
 import omit from 'lodash-es/omit';
-import { ResolveDesignValueType } from './useBreakpoints';
 import { getUnboundValues } from '@/utils/getUnboundValues';
 import { DropZone } from '@components/DropZone/Dropzone';
-
-import { Link } from 'contentful';
+import { useEntityStore } from '@/store/entityStore';
 
 type PropsType =
   | StyleProps
@@ -42,16 +43,12 @@ export const useComponentProps = ({
   resolveDesignValue,
   definition,
 }: ComponentPropsParams) => {
-  const { entityStore, unboundValues, dataSource } = useEditorStore(
-    useShallow((state) => ({
-      entityStore: state.entityStore,
-      unboundValues: state.unboundValues,
-      dataSource: state.dataSource,
-    }))
-  );
-
+  const unboundValues = useEditorStore((state) => state.unboundValues);
+  const dataSource = useEditorStore((state) => state.dataSource);
+  const entityStore = useEntityStore((state) => state.entityStore);
   const props: PropsType = useMemo(() => {
-    if (!definition) {
+    // Don't enrich the design component wrapper node with props
+    if (!definition || node.type === DESIGN_COMPONENT_NODE_TYPE) {
       return {};
     }
 
@@ -66,7 +63,10 @@ export const useComponentProps = ({
         }
 
         if (variableMapping.type === 'DesignValue') {
-          const valueByBreakpoint = resolveDesignValue(variableMapping.valuesByBreakpoint);
+          const valueByBreakpoint = resolveDesignValue(
+            variableMapping.valuesByBreakpoint,
+            variableName
+          );
           const designValue =
             variableName === 'cfHeight'
               ? calculateNodeDefaultHeight({
@@ -81,13 +81,34 @@ export const useComponentProps = ({
             [variableName]: designValue,
           };
         } else if (variableMapping.type === 'BoundValue') {
-          // take value from the datasource for both bound and unbound value types
+          // // take value from the datasource for both bound and unbound value types
           const [, uuid, ...path] = variableMapping.path.split('/');
-          const binding = dataSource[uuid];
+          const binding = dataSource[uuid] as Link<'Entry' | 'Asset'>;
 
-          const boundValue = areEntitiesFetched
-            ? entityStore?.getValue(binding, path.slice(0, -1))
+          let boundValue: string | Link<'Asset'> | undefined = areEntitiesFetched
+            ? entityStore.getValue(binding, path.slice(0, -1))
             : undefined;
+          // In some cases, there may be an asset linked in the path, so we need to consider this scenario:
+          // If no 'boundValue' is found, we also attempt to extract the value associated with the second-to-last item in the path.
+          // If successful, it means we have identified the linked asset.
+
+          if (!boundValue) {
+            const maybeBoundAsset = areEntitiesFetched
+              ? entityStore.getValue(binding, path.slice(0, -2))
+              : undefined;
+
+            if (isLinkToAsset(maybeBoundAsset)) {
+              boundValue = maybeBoundAsset;
+            }
+          }
+
+          if (
+            typeof boundValue === 'object' &&
+            (boundValue as Link<'Entry' | 'Asset'>).sys.linkType === 'Asset'
+          ) {
+            boundValue = entityStore?.getValue(boundValue, ['fields', 'file']);
+          }
+
           const value = boundValue || variableDefinition.defaultValue;
 
           return {
@@ -117,8 +138,9 @@ export const useComponentProps = ({
     resolveDesignValue,
     dataSource,
     areEntitiesFetched,
-    entityStore,
     unboundValues,
+    node.type,
+    entityStore,
   ]);
 
   const cfStyles = buildCfStyles(props);
@@ -146,14 +168,13 @@ export const useComponentProps = ({
 
   const { className } = useStyleTag({ styles: cfStyles, nodeId: node.data.id });
 
-  const renderDropZone = (node: CompositionComponentNode, props?: Record<string, any>) => {
+  const renderDropZone = (node: CompositionComponentNode, props?: Record<string, unknown>) => {
     return (
       <DropZone
         sectionId={node.data.id}
         zoneId={node.data.id}
         node={node}
         resolveDesignValue={resolveDesignValue}
-        areEntitiesFetched={areEntitiesFetched}
         {...props}
       />
     );
@@ -161,7 +182,7 @@ export const useComponentProps = ({
 
   const defaultedProps = {
     className,
-    editMode: true,
+    editorMode: true,
     node,
     renderDropZone,
     'data-cf-node-id': node.data.id,